refactor(router): fold elements-meta into router-state; document public APIs
Drop the thin elements-meta module and put Navigate settle docs on useRouter,
with brief JSDoc on Link and Router.
